--- app/views/App/Hello.html
@@ -1,7 +1,0 @@
-{{set . "title" "Hello page"}}
-{{template "header.html" .}}
-
-<h1>Hello {{.myName}}</h1>
-<a href="/">Back to form</a>
-
-{{template "footer.html" .}}
