--- app/views/footer.html
@@ -1,5 +1,0 @@
-  {{if .DevMode}}
-    {{template "debug.html" .}}
-  {{end}}
-  </body>
-</html>
