--- conf/routes
@@ -1,27 +1,0 @@
-# Routes Config
-#
-# This file defines all application routes (Higher priority routes first)
-#
-
-module:testrunner
-# module:jobs
-
-
-GET     /                                       App.Index
-GET     /App/Hello                              App.Hello
-
-# Ignore favicon requests
-GET     /favicon.ico                            404
-
-# Map static resources from the /app/public folder to the /public path
-GET     /public/*filepath                       Static.Serve("public")
-
-# Catch all, this will route any request into the controller path
-#
-#                    **** WARNING ****
-# Enabling this exposes any controller and function to the web.
-# ** This is a serious security issue if used online **
-#
-# For rapid development uncomment the following to add new controller.action endpoints
-# without having to add them to the routes table.
-# *       /:controller/:action                    :controller.:action
