--- go.mod
@@ -1,35 +1,0 @@
-module go_reval
-
-go 1.17
-
-require (
-	github.com/revel/modules v1.1.0
-	github.com/revel/revel v1.1.0
-)
-
-require (
-	github.com/agtorre/gocolorize v1.0.0 // indirect
-	github.com/bradfitz/gomemcache v0.0.0-20220106215444-fb4bf637b56d // indirect
-	github.com/fsnotify/fsnotify v1.5.1 // indirect
-	github.com/go-stack/stack v1.8.1 // indirect
-	github.com/gomodule/redigo v1.8.8 // indirect
-	github.com/google/uuid v1.3.0 // indirect
-	github.com/inconshreveable/log15 v0.0.0-20201112154412-8562bdadbbac // indirect
-	github.com/jessevdk/go-flags v1.4.0 // indirect
-	github.com/mattn/go-colorable v0.1.12 // indirect
-	github.com/mattn/go-isatty v0.0.14 // indirect
-	github.com/patrickmn/go-cache v2.1.0+incompatible // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	github.com/revel/cmd v1.1.2 // indirect
-	github.com/revel/config v1.1.0 // indirect
-	github.com/revel/log15 v2.11.20+incompatible // indirect
-	github.com/revel/pathtree v0.0.0-20140121041023-41257a1839e9 // indirect
-	github.com/xeonx/timeago v1.0.0-rc4 // indirect
-	golang.org/x/mod v0.6.0-dev.0.20220106191415-9b9b3d81d5e3 // indirect
-	golang.org/x/net v0.0.0-20220412020605-290c469a71a5 // indirect
-	golang.org/x/sys v0.0.0-20220227234510-4e6760a101f9 // indirect
-	golang.org/x/tools v0.1.10 // indirect
-	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
-	gopkg.in/natefinch/lumberjack.v2 v2.0.0 // indirect
-	gopkg.in/stack.v0 v0.0.0-20141108040640-9b43fcefddd0 // indirect
-)
