+++ myapp/app/views/App/Hello.html
@@ -1,0 +1,7 @@
+{{set . "title" "Hello page"}}
+{{template "header.html" .}}
+
+<h1>Hello {{.myName}}</h1>
+<a href="/">Back to form</a>
+
+{{template "footer.html" .}}
