+++ myapp/app/views/footer.html
@@ -1,0 +1,5 @@
+  {{if .DevMode}}
+    {{template "debug.html" .}}
+  {{end}}
+  </body>
+</html>
